requirements.txt
# LeafSense Requirements - Simplified
# ====================================

# Core Web Framework
Flask==3.0.3
Werkzeug==3.0.3

# Machine Learning
tensorflow>=2.20.0
numpy>=2.0.0

# Image Processing
Pillow>=10.0.0

# Environment Configuration
python-dotenv==1.0.1

# Production Server (optional)
gunicorn==23.0.0
